requirements.txt
astroid==2.4.2
attrs==19.3.0
Django==2.2.9
django-debug-toolbar==2.2
flake8==3.8.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.8.0
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
wcwidth==0.2.5
wrapt==1.12.1
